Refactor code to enable testing
